Vivaldi Algorithm, Initial Sched, Continuous Sched
[S] Adaptation Manager Initialization

Created low level functions to execute the adaptation manager. All
functions are executed and scheduled in the mk-supervisor function of
the supervisor.clj file

[S] Network Space (Vivaldi alg.)

Creation of the network space in a distributed manner, using the Vivaldi
algorithm.

Each node launch a server to let others know its coordinates and
prediction error. The algorithm it's periodically executed to recompute
coordinates according to measurements with other nodes registred into
the cluster.

Publish coordinates to ZK
	Save coordinates to ZooKeeper when they are considered
	stable (confidence is greater than a threshold).

Retrieve setting
	Parameters are retrieved from storm configuration file
	(which has been extended with needed information).

[S] Continuous Scheduler & WorkerMonitor

Low level invocation procedure

Clojure function to execute continuous scheduler. This is the low level
scheduler function which has to retrieve local topologies and nodes
alive into the cluster.

It's executed by the supervisor. The business logic of the scheduler has
to be implemented into the Continuous Scheduler classes.