Network Aware extension by UniRoma2
This branch includes an extension for Apache Storm, which it's able to
schedule topologies using a location-aware strategy.

Context: We want to use distributed nodes to execute topologies in world
wide distributed network.

Scheduling Criteria: The extention is configured to minimize the network
usage, defined as (inter-node traffic)*(inter-node latencies).

Author: Matteo Nardelli
University: University of Rome "Tor Vergata"